requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ibm>=1.0.0
ibm-watsonx-ai>=1.3.0

chromadb>=0.5.0
pypdf>=5.0.0

python-dotenv>=1.0.1
